chunk-audio-api/requirements.txt
fastapi
uvicorn
google-cloud-speech
google-generativeai
python-multipart
pymongo
python-dotenv